Renames 'field' into 'attributeName' for SimpleBooleanQuery, SimpleFulltextQuery and ProximityQuery
