+++ vitrivr-engine-plugin-features/src/main/kotlin/org/vitrivr/engine/base/features/external/implementations/dino/DINORetriever.kt
@@ -9,10 +9,9 @@ import org.vitrivr.engine.core.features.AbstractRetriever
 import org.vitrivr.engine.core.model.content.element.ImageContent
 import org.vitrivr.engine.core.model.descriptor.vector.FloatVectorDescriptor
 import org.vitrivr.engine.core.model.metamodel.Schema
-import org.vitrivr.engine.core.model.query.proximity.Distance
+import org.vitrivr.engine.core.model.query.basics.Distance
 import org.vitrivr.engine.core.model.query.proximity.ProximityQuery
 import org.vitrivr.engine.core.model.retrievable.Retrieved
-import org.vitrivr.engine.core.model.retrievable.attributes.DistanceAttribute
 import org.vitrivr.engine.core.model.retrievable.attributes.ScoreAttribute
 
 /**
@@ -32,14 +31,6 @@ class DINORetriever(
     query: FloatVectorDescriptor,
     context: QueryContext
 ) : AbstractRetriever<ImageContent, FloatVectorDescriptor>(field, query, context) {
-
-    companion object {
-        fun scoringFunction(retrieved: Retrieved): Float {
-            val distance = retrieved.filteredAttribute<DistanceAttribute>()?.distance ?: return 0f
-            return 1f - distance
-        }
-    }
-
     override fun toFlow(scope: CoroutineScope): Flow<Retrieved> {
         val k = context.getProperty(field.fieldName, "limit")?.toIntOrNull() ?: 1000 //TODO get limit
         val returnDescriptor = context.getProperty(field.fieldName, "returnDescriptor")?.toBooleanStrictOrNull() ?: false
